Replace optimist with yargs
Optimist is no longer maintained. It has not been updated since 2014,
even for basic security updates. Thus this change replaces it with its
spiritual successor, yargs, which is still actively maintained.

This is also a first step to removing peerjs' dependence on an outdated
version of minimist that has known security vulnerabilities.